curries_portal/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.16
django-admin-interface==0.20.0
django-admin-numeric-filter==0.1.6
django-admin-volt==0.0.5
django-adminlte-3==0.1.6
django-advanced-filters==2.0.0
django-allauth==0.51.0
django-autocompletefilter==0.0.8
django-bootstrap-v5==1.0.11
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.8.0
django-jazzmin==2.5.0
django-livesync==0.5
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-timezone-field==5.0
django-userforeignkey==0.5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.1
et-xmlfile==1.1.0
fpdf==1.7.2
idna==3.4
Markdown==3.4.1
MarkupPy==1.14
mysqlclient==2.1.1
oauthlib==3.2.1
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
pycparser==2.21
PyJWT==2.5.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
tornado==6.2
types-cryptography==3.3.23
tzdata==2022.2
urllib3==1.26.12
watchdog==2.1.9
websocket-client==1.4.1
xlrd==2.0.1
xlwt==1.3.0
